First script to run Prokka on genomes and genomes
This is the first script I've developed to run prokka on all megalocytivirus and iridoviridae genomes. Currently doesn't work and i need to fix it. Also being uploaded are the FASTA files for each megalocytivirus genome and iridoviridae genome i want to run prokka and other programs on as part of a bioinformaticv pipeline.