Updated paths relative to directory instructions in README
I updated the paths in the script relative to the instruction now added to the README file in the repository, asking the user to create a directory called mcv.

I also removed some of the commentary / notes in the code which are no longer applicable to the script (i.e. instructions for me with regard to the use of alternative file paths depending on where I'm running the script from - desktop vs server).