Deleting results as did not include genome AF303741
I noticed that genome AF303741 (an Iridovirus genome), was missing from the genomes included in the previous analysis (and therefore this output). I added the genome to the input (Genomes) folder and re-ran the analysis yesterday. The new outputs, which included genome AF303741, will replace the outputs I'm deleting from this folder.