Inserted code to remove trailing spaces and removed direction to use .sh file
I've done this as users' paths to their MultipleSequenceAlignment directory (an output of Orthofinder) will vary. This is as Orthofinder names the results directory by including the day and month the results are generated (for example, ResultsJul15). 

As such, I don't believe I can use a generic .sh file containing the code to remove trailing spaces, with a generic path included for the MultipleSequenceAlignment directory. Instead, users will have to read the code and insert their unique path to the MultipleSequenceAlignment directory themselves, which will include the unique name of their results directory.