Removed -n option as prthofinder script will always include a unique identifier in the results directory.
I added the -n option to specify a name for the results directory where outputs are saved from Orthofinder. This was to include the names of  the directory (which wouldn't change) in the following script. However, I now understand Orthofinder will always add a unique identified to the results directory, being the date, (Results_Jul15), regardless of whether the -n option is used or not.