Pushing iqtree results for gene trees
These are iqtree outputs from the first successful run of the script on this Github repository named 'script_four_iqtree.sh'. The multiple sequence alignments (MSAs) taken as input are the first 63 MSAs produced by Orthofinder earlier in my bioinformatic pipeline. These 63 MSAs are alignments of orthogroups which are representing in the majority of megalocytivirus genomes.