Updates READ.ME to include iqtree script instructions and instructions to sort iqtree output files.
This included instructions at the beginning of the script to add the parallel module to ones Conda environment, in order to invoke it as part of the iqtree script later down the pipeline. I need to test the Conda command to ensure it works, as I have been using an HPC with the module already loaded.