Replacing the multiple sequence alignment files with sequence names updated, with an earlier version of each file without sequence names updated.
I have done this is I have realised that the script used to update the sequence names has changed the sequences themselves. This should not be happening. Until I fix the big in the script which updates the sequence names, I have replaced the multiple sequence alignment files with the .fa files which do not have the sequence names updated.