Add script to make gene trees
Note that this script uses QMaker to make two models:

1. For iridoviridae alignemnts
2. For mcv alignments

I don't make one for the pagrus1, becuase likely there's not enough subsitutions to do it well, and in addition it's very likely to be very very similar to the mcv model.

Note - we should not use or believe these models (or the trees they are used to estimate) without first checking the model parameters themselves. I will add a script to do this too, but essentially we compare them to existing models like WAG and LG, and examine the actual parameters to make sure they're not crazy and/or at the limits (like 0 or 100).