rename script qmaker
Makes it clear that this script is not the final say on the gene trees. Rather, we're just getting the models. Here we can view the gene trees we estimate as a kind of QC on the models themselves: most of the iridoviridae genes should prefer Q.iridoviridae, and most of the MCV genes should prefer Q.mcv.

We need to do additional QC, like looking at the model files and doing some sanity-checking of the parameters, PCAs, etc.